Convert more tests in input.spec.jsx
* Rename input.spec.js to input.spec.jsx
